+++ contracts/ethStaking/interfaces/IRewardsReceiver.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.15;
 
 interface IRewardsReceiver {
-    function withdraw() external returns (uint);
+    function withdraw() external returns (uint256);
 }
